made result id field an int
Former-commit-id: 2cf08abeb314b28ee66ece373c8f586fb6984082

+++ src/mitll/langtest/client/LangTestDatabase.java
@@ -182,7 +182,7 @@ public interface LangTestDatabase extends RemoteService {
    * @param resultid
    * @param roundTrip
    */
-  void addRoundTrip(long resultid, int roundTrip);
+  void addRoundTrip(int resultid, int roundTrip);
 
   /**
    * @see mitll.langtest.client.scoring.SimplePostAudioRecordButton#postAudioFile(String)
